src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
boto3==1.26.73
    # via -r requirements.in
botocore==1.29.73
    # via
    #   boto3
    #   s3transfer
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
numpy==1.24.2
    # via -r requirements.in
pillow==9.4.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via botocore
pyyaml==6.0
    # via -r requirements.in
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via python-dateutil
urllib3==1.26.14
    # via botocore
